refactor: 주소 변환 메서드명 변경
