+++ mealkitary-infrastructure/adapter-address-resolver/simple-address-resolver/src/test/kotlin/com/mealkitary/address/SimpleAddressResolverTest.kt
@@ -14,7 +14,7 @@ class SimpleAddressResolverTest : AnnotationSpec() {
         val address = "서울특별시 강남구 역삼동 논현로"
         val resolver = SimpleAddressResolver()
 
-        val shopAddress = resolver.resolveAddress(address)
+        val shopAddress = resolver.resolve(address)
 
         shopAddress.cityCode shouldBe "1234567890"
         shopAddress.coordinates.longitude shouldBe "127.0"
@@ -30,7 +30,7 @@ class SimpleAddressResolverTest : AnnotationSpec() {
         val address = "경기도 남양주시 다산동"
         val resolver = SimpleAddressResolver()
 
-        val shopAddress = resolver.resolveAddress(address)
+        val shopAddress = resolver.resolve(address)
 
         shopAddress.cityCode shouldBe "1234567890"
         shopAddress.coordinates.longitude shouldBe "127.0"
@@ -46,7 +46,7 @@ class SimpleAddressResolverTest : AnnotationSpec() {
         val address = "제주특별자치도 한림읍"
         val resolver = SimpleAddressResolver()
 
-        val shopAddress = resolver.resolveAddress(address)
+        val shopAddress = resolver.resolve(address)
 
         shopAddress.cityCode shouldBe "1234567890"
         shopAddress.coordinates.longitude shouldBe "127.0"
@@ -63,7 +63,7 @@ class SimpleAddressResolverTest : AnnotationSpec() {
         val resolver = SimpleAddressResolver()
 
         shouldThrow<IllegalArgumentException> {
-            resolver.resolveAddress(address)
+            resolver.resolve(address)
         } shouldHaveMessage "주소 형식이 올바르지 않습니다."
     }
 }
